+++ src/pgeditor/gui/graph/node_style.h
@@ -16,14 +16,14 @@ namespace pgeditor::gui::graph
 
     namespace node
     {
-      const QColor border_color = QColor(0,0,0);
+      const QColor border_color = QColor(235,219,178);
       const uint32_t border_width = 2;
-      const uint32_t round_size = 5;
+      const uint32_t round_size = 0;
 
-      const QColor fill_gradient_start = QColor(0,52,255);
-      const QColor fill_gradient_end = QColor(0,40,192);
-      const QColor selected_fill_gradient_start = QColor(255,0,0);
-      const QColor selected_fill_gradient_end = QColor(156,0,0);
+      const QColor fill_gradient_start = QColor(60,56,54);
+      const QColor fill_gradient_end = QColor(60,56,54);
+      const QColor selected_fill_gradient_start = QColor(146,131,116);
+      const QColor selected_fill_gradient_end = QColor(146,131,116);
 
       const uint32_t ports_top_margin = 3;
       const uint32_t ports_bottom_margin = 3;
@@ -37,21 +37,21 @@ namespace pgeditor::gui::graph
       const uint32_t horizontal_spacing = 3;
       const char* const name_font_family = "Helvetica";
       const uint32_t name_font_size = 10;
-      const char* const label_stylesheet = "background: transparent; color: white";
+      const char* const label_stylesheet = "background: transparent;";
     }
 
     namespace port_connection
     {
       const QColor border_color = QColor(0, 0, 0);
       const uint32_t border_width = 2;
-      const QColor fill_color = QColor(255,255,255);
+      const QColor fill_color = QColor(235,219,178);
       const uint32_t port_size = 10;
     }
 
     namespace connector
     {
-      const QColor line_color = QColor(0, 0, 0);
-      const uint32_t line_width = 2;
+      const QColor line_color = QColor(235,219,178);
+      const uint32_t line_width = 1;
       const uint32_t control_point_distance = 50;
     }